+++ tests/schemas/test_ase.py
@@ -103,16 +103,18 @@ def test_summarize_opt_run():
     atoms = bulk("Cu") * (2, 2, 1)
     atoms[0].position += [0.1, 0.1, 0.1]
     atoms.calc = EMT()
-    BFGS(atoms, trajectory="test.traj").run()
+    dyn = BFGS(atoms, trajectory="test.traj")
+    dyn.trajectory = "test.traj"  # can remove after ASE MR 2901
+    dyn.run()
     traj = read("test.traj", index=":")
 
-    results = summarize_opt_run(traj)
+    results = summarize_opt_run(dyn, check_convergence=True)
     assert results["nsites"] == len(atoms)
     assert results["atoms"] == traj[-1]
     assert results["results"]["energy"] == atoms.get_potential_energy()
     assert len(results["trajectory"]) == len(traj)
     assert len(results["trajectory_results"]) == len(traj)
-    assert results["trajectory_results"][-1] == results["results"]
+    assert results["trajectory_results"][-1]["energy"] == results["results"]["energy"]
     assert "nid" in results
     assert "dir_name" in results
     assert "pull_request" in results["builder_meta"]
@@ -121,16 +123,18 @@ def test_summarize_opt_run():
     atoms = bulk("Cu") * (2, 2, 1)
     atoms[0].position += [0.1, 0.1, 0.1]
     atoms.calc = EMT()
-    BFGS(atoms, trajectory="test.traj").run()
+    dyn = BFGS(atoms, trajectory="test.traj")
+    dyn.trajectory = "test.traj"  # can remove after ASE MR 2901
+    dyn.run()
     traj = read("test.traj", index=":")
 
-    results = summarize_opt_run(traj, remove_empties=True)
+    results = summarize_opt_run(dyn, remove_empties=True)
     assert results["nsites"] == len(atoms)
     assert results["atoms"] == traj[-1]
     assert results["results"]["energy"] == atoms.get_potential_energy()
     assert len(results["trajectory"]) == len(traj)
     assert len(results["trajectory_results"]) == len(traj)
-    assert results["trajectory_results"][-1] == results["results"]
+    assert results["trajectory_results"][-1]["energy"] == results["results"]["energy"]
     assert "nid" in results
     assert "dir_name" in results
     assert "pull_request" not in results["builder_meta"]
@@ -140,49 +144,15 @@ def test_summarize_opt_run():
     atoms.info["test_dict"] = {"hi": "there", "foo": "bar"}
     atoms[0].position += [0.1, 0.1, 0.1]
     atoms.calc = EMT()
-    BFGS(atoms, trajectory="test.traj").run()
+    dyn = BFGS(atoms, trajectory="test.traj")
+    dyn.trajectory = "test.traj"  # can remove after ASE MR 2901
+    dyn.run()
     traj = read("test.traj", index=":")
 
-    results = summarize_opt_run(traj)
+    results = summarize_opt_run(dyn)
     assert results.get("atoms_info", {}) != {}
     assert results["atoms_info"].get("test_dict", None) == {"hi": "there", "foo": "bar"}
     assert results["atoms"].info.get("test_dict", None) == {"hi": "there", "foo": "bar"}
-
-    # Make sure magnetic moments are handled appropriately
-    atoms = bulk("Cu") * (2, 2, 1)
-    atoms.set_initial_magnetic_moments([3.14] * len(atoms))
-    atoms.info["test_dict"] = {"hi": "there", "foo": "bar"}
-    atoms[0].position += [0.1, 0.1, 0.1]
-    atoms.calc = EMT()
-    BFGS(atoms, trajectory="test.traj").run()
-    traj = read("test.traj", index=":")
-    traj[-1].calc.results["magmoms"] = [2.0] * len(atoms)
-
-    results = summarize_opt_run(traj)
-
-    assert atoms.calc is not None
-    assert atoms.get_initial_magnetic_moments().tolist() == [3.14] * len(atoms)
-
-    assert results["atoms"].get_initial_magnetic_moments().tolist() == [2.0] * len(
-        atoms
-    )
-    assert results["atoms"].calc is None
-
-    # Make sure Atoms magmoms were not moved if specified
-    atoms = bulk("Cu") * (2, 2, 1)
-    atoms.set_initial_magnetic_moments([3.14] * len(atoms))
-    atoms.info["test_dict"] = {"hi": "there", "foo": "bar"}
-    atoms[0].position += [0.1, 0.1, 0.1]
-    atoms.calc = EMT()
-    BFGS(atoms, trajectory="test.traj").run()
-    traj = read("test.traj", index=":")
-    traj[-1].calc.results["magmoms"] = [2.0] * len(atoms)
-
-    results = summarize_opt_run(traj, prep_next_run=False)
-    assert atoms.get_initial_magnetic_moments().tolist() == [3.14] * len(atoms)
-    assert results["atoms"].get_initial_magnetic_moments().tolist() == [3.14] * len(
-        atoms
-    )
 
     # test document can be jsanitized and decoded
     d = jsanitize(results, strict=True, enum_values=True)